teat: 测试

+++ src/main/java/org/chenjh/aiqasystem/repo/JOOQRepository.java
@@ -15,7 +15,7 @@ public interface JOOQRepository<T> {
 
     List<T> findAll();
 
-    Optional<T> findById(int id);
+    Optional<T> findById(long id);
 
     boolean deleteById(int id);
 }
